extract KeystrokeInput into it's own file
