+++ crates/settings_ui/src/ui_components/mod.rs
@@ -1,1 +1,2 @@
+pub mod keystroke_input;
 pub mod table;
